[Doc] Optimize sidebar nav creation to speed up build
