+++ doc/source/custom_directives.py
@@ -1,24 +1,32 @@
+import re
+import sphinx
+from typing import List, Dict, Union, Callable, Any
+import copy
+import yaml
+import bs4
 import logging
 import logging.handlers
+import pathlib
 import urllib
 import urllib.request
-from pathlib import Path
 from queue import Queue
-from urllib.parse import urlparse
-import yaml
 import requests
 from preprocess_github_markdown import preprocess_github_markdown_file
 from sphinx.util import logging as sphinx_logging
 from sphinx.util.console import red  # type: ignore
+from sphinx.util.nodes import make_refnode
+from docutils import nodes
+from functools import lru_cache
+from pydata_sphinx_theme.toctree import add_collapse_checkboxes
+
+logger = logging.getLogger(__name__)
 
 __all__ = [
     "DownloadAndPreprocessEcosystemDocs",
     "update_context",
     "LinkcheckSummarizer",
-    "build_gallery",
+    "setup_context",
 ]
-
-GALLERIES = ["ray-overview/eco-gallery.yml"]
 
 # Taken from https://github.com/edx/edx-documentation
 FEEDBACK_FORM_FMT = (
@@ -64,7 +72,7 @@ class DownloadAndPreprocessEcosystemDocs:
     """
 
     def __init__(self, base_path: str) -> None:
-        self.base_path = Path(base_path).absolute()
+        self.base_path = pathlib.Path(base_path).absolute()
         assert self.base_path.is_dir()
         self.original_docs = {}
 
@@ -179,64 +187,239 @@ class LinkcheckSummarizer:
             self.logger.info("No broken links found!")
 
 
-def build_gallery(app):
-    for gallery in GALLERIES:
-        panel_items = []
-        source = yaml.safe_load((Path(app.srcdir) / gallery).read_text())
+def parse_navbar_config(app: sphinx.application.Sphinx, config: sphinx.config.Config):
+    """Parse the navbar config file into a set of links to show in the navbar.
 
-        meta = source["meta"]
-        grid = meta.pop("grid")
-        projects = source["projects"]
-        classes = source["classes"]
+    Parameters
+    ----------
+    app : sphinx.application.Sphinx
+        Application instance passed when the `config-inited` event is emitted
+    config : sphinx.config.Config
+        Initialized configuration to be modified
+    """
+    if "navbar_content_path" in config:
+        filename = app.config["navbar_content_path"]
+    else:
+        filename = ""
 
-        for item in projects:
-            ref = "button-link"
-            website = item["website"]
-            if "://" not in website:  # if it has no http/s protocol, it's a "ref"
-                ref = ref.replace("link", "ref")
+    if filename:
+        with open(pathlib.Path(__file__).parent / filename, "r") as f:
+            config.navbar_content = yaml.safe_load(f)
+    else:
+        config.navbar_content = None
 
-            if not item.get("image"):
-                item["image"] = "https://docs.ray.io/_images/ray_logo.png"
-            gh_stars = ""
-            if item["repo"]:
-                try:
-                    url = urlparse(item["repo"])
-                    if url.netloc == "github.com":
-                        _, org, repo = url.path.rstrip("/").split("/")
-                        gh_stars = (
-                            f".. image:: https://img.shields.io/github/"
-                            f"stars/{org}/{repo}?style=social)]\n"
-                            f"\t\t\t:target: {item['repo']}"
-                        )
-                except Exception:
-                    pass
 
-            item = f"""
-    .. grid-item-card::
-        :img-top: {item["image"]}
-        :class-img-top: {classes["class-img-top"]}
+NavEntry = Dict[str, Union[str, List["NavEntry"]]]
 
-        {gh_stars}
 
-        {item["description"]}
+def preload_sidebar_nav(get_toctree: Callable[[Any], str]) -> bs4.BeautifulSoup:
+    """Return the navigation link structure in HTML.
 
-        +++
-        .. {ref}:: {item["website"]}
-            :color: primary
-            :outline:
-            :expand:
+    This function is modified from the pydata_sphinx_theme function
+    `generate_toctree_html`. However, for ray we only produce one
+    sidebar for all pages. We therefore can call this function just once (per worker),
+    cache the result, and reuse it.
 
-            {item["name"]}
-            """
+    The result of this function is equivalent to calling
 
-            panel_items.append(item)
+        pydata_sphinx_theme.toctree.generate_toctree_html(
+            "sidebar",
+            startdepth=0,
+            show_nav_level=0,
+            collapse=False,
+            includehidden=True,
+            maxdepth=4,
+            titles_only=True
+        )
 
-        panel_header = f".. grid:: {grid}\n"
-        for k, v in meta.items():
-            panel_header += f"    :{k}: {v}\n"
+    Here we cache the result on this function itself because the `get_toctree`
+    function is not the same instance for each `html-page-context` event, even
+    if the toctree content is identical.
 
-        panel_items = "\n".join(panel_items)
-        panels = panel_header + panel_items
+    Parameters
+    ----------
+    get_toctree : Callable[[Any], str]
+        The function defined in context["toctree"] when html-page-context is triggered
 
-        gallery_out = gallery.replace(".yml", ".txt")
-        (Path(app.srcdir) / gallery_out).write_text(panels)
+    Returns
+    -------
+    bs4.BeautifulSoup
+        Soup to display in the side navbar
+    """
+    if hasattr(preload_sidebar_nav, "cached_toctree"):
+        return preload_sidebar_nav.cached_toctree
+
+    toctree = get_toctree(
+        collapse=False, includehidden=True, maxdepth=4, titles_only=True
+    )
+
+    soup = bs4.BeautifulSoup(toctree, "html.parser")
+
+    # All the URIs are relative to the current document location, e.g.
+    # "../../<whatever.html>". There's no need for this, and it messes up the build if
+    # we reuse the same sidebar (with different relative URIs for various pages) on
+    # different pages. Replace the leading "../" sequences in the hrefs
+    for a in soup.select("a"):
+        a["href"] = re.sub(r"^(\.\.\/)*", "/", a["href"])
+
+    # pair "current" with "active" since that's what we use w/ bootstrap
+    for li in soup("li", {"class": "current"}):
+        li["class"].append("active")
+
+    # Remove sidebar links to sub-headers on the page
+    for li in soup.select("li"):
+        # Remove
+        if li.find("a"):
+            href = li.find("a")["href"]
+            if "#" in href and href != "#":
+                li.decompose()
+
+    # Add bootstrap classes for first `ul` items
+    for ul in soup("ul", recursive=False):
+        ul.attrs["class"] = ul.attrs.get("class", []) + ["nav", "bd-sidenav"]
+
+    # Add collapse boxes for parts/captions.
+    # Wraps the TOC part in an extra <ul> to behave like chapters with toggles
+    # show_nav_level: 0 means make parts collapsible.
+    partcaptions = soup.find_all("p", attrs={"class": "caption"})
+    if len(partcaptions):
+        new_soup = bs4.BeautifulSoup("<ul class='list-caption'></ul>", "html.parser")
+        for caption in partcaptions:
+            # Assume that the next <ul> element is the TOC list
+            # for this part
+            for sibling in caption.next_siblings:
+                if sibling.name == "ul":
+                    toclist = sibling
+                    break
+            li = soup.new_tag("li", attrs={"class": "toctree-l0"})
+            li.extend([caption, toclist])
+            new_soup.ul.append(li)
+        soup = new_soup
+
+    # Add icons and labels for collapsible nested sections
+    add_collapse_checkboxes(soup)
+
+    preload_sidebar_nav.cached_toctree = soup
+    return soup
+
+
+def setup_context(app, pagename, templatename, context, doctree):
+    @lru_cache(maxsize=None)
+    def render_header_nav_links() -> bs4.BeautifulSoup:
+        """Render external header links into the top nav bar.
+        The structure rendered here is defined in an external yaml file.
+
+        Returns
+        -------
+        str
+            Raw HTML to be rendered in the top nav bar
+        """
+        if not hasattr(app.config, "navbar_content"):
+            raise ValueError(
+                "A template is attempting to call render_header_nav_links(); a "
+                "navbar configuration must be specified."
+            )
+
+        node = nodes.container(classes=["navbar-content"])
+        node.append(render_header_nodes(app.config.navbar_content))
+        header_soup = bs4.BeautifulSoup(
+            app.builder.render_partial(node)["fragment"], "html.parser"
+        )
+        return add_nav_chevrons(header_soup)
+
+    def render_header_nodes(
+        obj: List[NavEntry], is_top_level: bool = True
+    ) -> nodes.Node:
+        """Generate a set of header nav links with docutils nodes.
+
+        Parameters
+        ----------
+        is_top_level : bool
+            True if the call to this function is rendering the top level nodes,
+            False otherwise (non-top level nodes are displayed as submenus of the top
+            level nodes)
+        obj : List[NavEntry]
+            List of yaml config entries to render as docutils nodes
+
+        Returns
+        -------
+        nodes.Node
+            Bullet list which will be turned into header nav HTML by the sphinx builder
+        """
+        bullet_list = nodes.bullet_list(
+            bullet="-",
+            classes=["navbar-toplevel" if is_top_level else "navbar-sublevel"],
+        )
+
+        for item in obj:
+
+            if "file" in item:
+                ref_node = make_refnode(
+                    app.builder,
+                    context["current_page_name"],
+                    item["file"],
+                    None,
+                    nodes.inline(classes=["navbar-link-title"], text=item.get("title")),
+                    item.get("title"),
+                )
+            elif "link" in item:
+                ref_node = nodes.reference("", "", internal=False)
+                ref_node["refuri"] = item.get("link")
+                ref_node["reftitle"] = item.get("title")
+                ref_node.append(
+                    nodes.inline(classes=["navbar-link-title"], text=item.get("title"))
+                )
+
+            if "caption" in item:
+                caption = nodes.Text(item.get("caption"))
+                ref_node.append(caption)
+
+            paragraph = nodes.paragraph()
+            paragraph.append(ref_node)
+
+            container = nodes.container(classes=["ref-container"])
+            container.append(paragraph)
+
+            list_item = nodes.list_item(
+                classes=["active-link"] if item.get("file") == pagename else []
+            )
+            list_item.append(container)
+
+            if "sections" in item:
+                wrapper = nodes.container(classes=["navbar-dropdown"])
+                wrapper.append(
+                    render_header_nodes(item["sections"], is_top_level=False)
+                )
+                list_item.append(wrapper)
+
+            bullet_list.append(list_item)
+
+        return bullet_list
+
+    context["cached_toctree"] = preload_sidebar_nav(context["toctree"])
+    context["render_header_nav_links"] = render_header_nav_links
+
+
+def add_nav_chevrons(input_soup: bs4.BeautifulSoup) -> bs4.BeautifulSoup:
+    """Add dropdown chevron icons to the header nav bar.
+
+    Parameters
+    ----------
+    input_soup : bs4.BeautifulSoup
+        Soup containing rendered HTML which will be inserted into the header nav bar
+
+    Returns
+    -------
+    bs4.BeautifulSoup
+        A new BeautifulSoup instance containing chevrons on the list items that
+        are meant to be dropdowns.
+    """
+    soup = copy.copy(input_soup)
+
+    for li in soup.find_all("li", recursive=True):
+        divs = li.find_all("div", {"class": "navbar-dropdown"}, recursive=False)
+        if divs:
+            ref = li.find("div", {"class": "ref-container"})
+            ref.append(soup.new_tag("i", attrs={"class": "fa-solid fa-chevron-down"}))
+
+    return soup
